+++ hutool-poi/src/main/java/cn/hutool/poi/excel/ExcelWriter.java
@@ -16,6 +16,7 @@ import cn.hutool.core.util.URLUtil;
 import cn.hutool.poi.excel.cell.CellLocation;
 import cn.hutool.poi.excel.cell.CellUtil;
 import cn.hutool.poi.excel.style.Align;
+import org.apache.poi.common.usermodel.Hyperlink;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.ss.usermodel.DataValidation;
@@ -947,6 +948,9 @@ public class ExcelWriter extends ExcelBase<ExcelWriter> {
 			} else {
 				rowMap = (Map) rowBean;
 			}
+		} else if (rowBean instanceof Hyperlink) {
+			// Hyperlink当成一个值
+			return writeRow(CollUtil.newArrayList(rowBean), isWriteKeyAsHead);
 		} else if (BeanUtil.isBean(rowBean.getClass())) {
 			if (MapUtil.isEmpty(this.headerAlias)) {
 				rowMap = BeanUtil.beanToMap(rowBean, new LinkedHashMap<>(), false, false);