+++ hutool-poi/src/main/java/cn/hutool/poi/excel/StyleSet.java
@@ -1,6 +1,7 @@
 package cn.hutool.poi.excel;
 
 import cn.hutool.poi.excel.style.StyleUtil;
+import org.apache.poi.hssf.util.HSSFColor;
 import org.apache.poi.ss.usermodel.BorderStyle;
 import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.ss.usermodel.FillPatternType;
@@ -14,7 +15,7 @@ import java.io.Serializable;
 
 /**
  * 样式集合，此样式集合汇集了整个工作簿的样式，用于减少样式的创建和冗余
- * 
+ *
  * @author looly
  *
  */
@@ -31,10 +32,12 @@ public class StyleSet implements Serializable{
 	protected CellStyle cellStyleForNumber;
 	/** 默认日期样式 */
 	protected CellStyle cellStyleForDate;
+	/** 默认链接样式 */
+	protected CellStyle cellStyleForHyperlink;
 
 	/**
 	 * 构造
-	 * 
+	 *
 	 * @param workbook 工作簿
 	 */
 	public StyleSet(Workbook workbook) {
@@ -42,56 +45,74 @@ public class StyleSet implements Serializable{
 		this.headCellStyle = StyleUtil.createHeadCellStyle(workbook);
 		this.cellStyle = StyleUtil.createDefaultCellStyle(workbook);
 
+		// 默认数字格式
+		cellStyleForNumber = StyleUtil.cloneCellStyle(workbook, this.cellStyle);
+		// 2表示：0.00
+		cellStyleForNumber.setDataFormat((short) 2);
+
 		// 默认日期格式
 		this.cellStyleForDate = StyleUtil.cloneCellStyle(workbook, this.cellStyle);
 		// 22表示：m/d/yy h:mm
 		this.cellStyleForDate.setDataFormat((short) 22);
 
-		// 默认数字格式
-		cellStyleForNumber = StyleUtil.cloneCellStyle(workbook, this.cellStyle);
-		// 2表示：0.00
-		cellStyleForNumber.setDataFormat((short) 2);
+
+		// 默认链接样式
+		this.cellStyleForHyperlink = StyleUtil.cloneCellStyle(workbook, this.cellStyle);
+		final Font font = this.workbook.createFont();
+		font.setUnderline((byte) 1);
+		font.setColor(HSSFColor.HSSFColorPredefined.BLUE.getIndex());
+		this.cellStyleForHyperlink.setFont(font);
 	}
 
 	/**
 	 * 获取头部样式，获取后可以定义整体头部样式
-	 * 
+	 *
 	 * @return 头部样式
 	 */
 	public CellStyle getHeadCellStyle() {
-		return headCellStyle;
+		return this.headCellStyle;
 	}
 
 	/**
 	 * 获取常规单元格样式，获取后可以定义整体头部样式
-	 * 
+	 *
 	 * @return 常规单元格样式
 	 */
 	public CellStyle getCellStyle() {
-		return cellStyle;
+		return this.cellStyle;
 	}
 
 	/**
-	 * 获取数字（带小数点）单元格样式，获取后可以定义整体头部样式
-	 * 
+	 * 获取数字（带小数点）单元格样式，获取后可以定义整体数字样式
+	 *
 	 * @return 数字（带小数点）单元格样式
 	 */
 	public CellStyle getCellStyleForNumber() {
-		return cellStyleForNumber;
+		return this.cellStyleForNumber;
 	}
 
 	/**
-	 * 获取日期单元格样式，获取后可以定义整体头部样式
-	 * 
+	 * 获取日期单元格样式，获取后可以定义整体日期样式
+	 *
 	 * @return 日期单元格样式
 	 */
 	public CellStyle getCellStyleForDate() {
-		return cellStyleForDate;
+		return this.cellStyleForDate;
+	}
+
+	/**
+	 * 获取链接单元格样式，获取后可以定义整体链接样式
+	 *
+	 * @return 链接单元格样式
+	 * @since 5.7.13
+	 */
+	public CellStyle getCellStyleForHyperlink() {
+		return this.cellStyleForHyperlink;
 	}
 
 	/**
 	 * 定义所有单元格的边框类型
-	 * 
+	 *
 	 * @param borderSize 边框粗细{@link BorderStyle}枚举
 	 * @param colorIndex 颜色的short值
 	 * @return this
@@ -102,12 +123,13 @@ public class StyleSet implements Serializable{
 		StyleUtil.setBorder(this.cellStyle, borderSize, colorIndex);
 		StyleUtil.setBorder(this.cellStyleForNumber, borderSize, colorIndex);
 		StyleUtil.setBorder(this.cellStyleForDate, borderSize, colorIndex);
+		StyleUtil.setBorder(this.cellStyleForHyperlink, borderSize, colorIndex);
 		return this;
 	}
 
 	/**
 	 * 设置cell文本对齐样式
-	 * 
+	 *
 	 * @param halign 横向位置
 	 * @param valign 纵向位置
 	 * @return this
@@ -118,12 +140,13 @@ public class StyleSet implements Serializable{
 		StyleUtil.setAlign(this.cellStyle, halign, valign);
 		StyleUtil.setAlign(this.cellStyleForNumber, halign, valign);
 		StyleUtil.setAlign(this.cellStyleForDate, halign, valign);
+		StyleUtil.setAlign(this.cellStyleForHyperlink, halign, valign);
 		return this;
 	}
 
 	/**
 	 * 设置单元格背景样式
-	 * 
+	 *
 	 * @param backgroundColor 背景色
 	 * @param withHeadCell 是否也定义头部样式
 	 * @return this
@@ -136,12 +159,13 @@ public class StyleSet implements Serializable{
 		StyleUtil.setColor(this.cellStyle, backgroundColor, FillPatternType.SOLID_FOREGROUND);
 		StyleUtil.setColor(this.cellStyleForNumber, backgroundColor, FillPatternType.SOLID_FOREGROUND);
 		StyleUtil.setColor(this.cellStyleForDate, backgroundColor, FillPatternType.SOLID_FOREGROUND);
+		StyleUtil.setColor(this.cellStyleForHyperlink, backgroundColor, FillPatternType.SOLID_FOREGROUND);
 		return this;
 	}
 
 	/**
 	 * 设置全局字体
-	 * 
+	 *
 	 * @param color 字体颜色
 	 * @param fontSize 字体大小，-1表示默认大小
 	 * @param fontName 字体名，null表示默认字体
@@ -155,7 +179,7 @@ public class StyleSet implements Serializable{
 
 	/**
 	 * 设置全局字体
-	 * 
+	 *
 	 * @param font 字体，可以通过{@link StyleUtil#createFont(Workbook, short, short, String)}创建
 	 * @param ignoreHead 是否跳过头部样式
 	 * @return this
@@ -168,12 +192,13 @@ public class StyleSet implements Serializable{
 		this.cellStyle.setFont(font);
 		this.cellStyleForNumber.setFont(font);
 		this.cellStyleForDate.setFont(font);
+		this.cellStyleForHyperlink.setFont(font);
 		return this;
 	}
 
 	/**
 	 * 设置单元格文本自动换行
-	 * 
+	 *
 	 * @return this
 	 * @since 4.5.16
 	 */
@@ -181,6 +206,8 @@ public class StyleSet implements Serializable{
 		this.cellStyle.setWrapText(true);
 		this.cellStyleForNumber.setWrapText(true);
 		this.cellStyleForDate.setWrapText(true);
+		this.cellStyleForHyperlink.setWrapText(true);
 		return this;
 	}
+
 }